+++ apps/common/src/main/resources/db/migration/canton-network/postgres/stable/V002__rename_ans_executed.sql
@@ -1,0 +1,1 @@
+ALTER TABLE dso_txlog_store RENAME COLUMN executed TO accepted;