+++ apps/sv/frontend/src/__tests__/mocks/handlers/sv-api.ts
@@ -1,7 +1,12 @@
 import { RestHandler, rest } from 'msw';
-import { ErrorResponse, GetDsoInfoResponse, ListDsoRulesVoteRequestsResponse } from 'sv-openapi';
+import {
+  ErrorResponse,
+  GetDsoInfoResponse,
+  ListDsoRulesVoteRequestsResponse,
+  ListDsoRulesVoteResultsResponse,
+} from 'sv-openapi';
 
-import { dsoInfo } from '../constants';
+import { dsoInfo, voteResults } from '../constants';
 
 export const buildSvMock = (svUrl: string): RestHandler[] => [
   rest.get(`${svUrl}/v0/admin/authorization`, (_, res, ctx) => {
@@ -16,6 +21,10 @@ export const buildSvMock = (svUrl: string): RestHandler[] => [
         dso_rules_vote_requests: [],
       })
     );
+  }),
+  rest.post(`${svUrl}/v0/admin/sv/voteresults`, (_, res, ctx) => {
+    console.log(voteResults);
+    return res(ctx.json<ListDsoRulesVoteResultsResponse>(voteResults));
   }),
   rest.get(`${svUrl}/v0/admin/domain/cometbft/debug`, (_, res, ctx) => {
     return res(