+++ apps/sv/frontend/src/contexts/SvContext.tsx
@@ -3,6 +3,7 @@ import { useSvClient } from 'common-frontend';
 import { Contract, PollingStrategy } from 'common-frontend-utils';
 
 import { AmuletRules } from '@daml.js/splice-amulet/lib/Splice/AmuletRules';
+import { SvNodeState } from '@daml.js/splice-dso-governance/lib/Splice/DSO/SvState/module';
 import { ElectionRequest, DsoRules } from '@daml.js/splice-dso-governance/lib/Splice/DsoRules';
 
 import { useSvAdminClient } from './SvAdminServiceContext';
@@ -15,6 +16,7 @@ type SvUiState =
       votingThreshold: bigint;
       amuletRules: Contract<AmuletRules>;
       dsoRules: Contract<DsoRules>;
+      nodeStatus: Contract<SvNodeState>;
     }
   | undefined;
 
@@ -32,6 +34,7 @@ export const useDsoInfos = (): UseQueryResult<SvUiState> => {
         votingThreshold: resp.voting_threshold,
         amuletRules: Contract.decodeOpenAPI(resp.amulet_rules, AmuletRules),
         dsoRules: Contract.decodeOpenAPI(resp.dso_rules, DsoRules),
+        nodeStatus: resp.sv_node_states.map(c => Contract.decodeOpenAPI(c, SvNodeState)),
       };
     },
   });