+++ apps/sv/src/main/scala/com/daml/network/sv/admin/http/HttpSvAdminHandler.scala
@@ -274,7 +274,7 @@ class HttpSvAdminHandler(
       for {
         voteResults <- dsoStore.listVoteRequestResults(
           body.actionName,
-          body.executed,
+          body.accepted,
           body.requester,
           body.effectiveFrom,
           body.effectiveTo,